Fixed CI error with accessing size as int instead of size_t

+++ onnx/version_converter/adapters/add_7_6.h
@@ -44,8 +44,8 @@ struct Add_7_6 final : public Adapter {
       // Determine what the broadcast dimension is - if necessary
       // Unnecessary if 1) all inputs are 1, 2) B is empty, 3) dimensions match
       // in reverse (assume this is required for model to compile in the first place?)
-      int axis = A_sizes.size() - B_sizes.size();
-      for (int i = B_sizes.size() - 1; i >= 0; i--) {
+      size_t axis = A_sizes.size() - B_sizes.size();
+      for (size_t i = B_sizes.size() - 1; i >= 0; i--) {
         ONNX_ASSERTM(axis >= 0, "Inputs are not broadcastable: no positive axis "
           "found to align dimensions.");
         if (B_sizes[i].dim == A_sizes[axis + i].dim || B_sizes[i].dim == 1) {
